rename module and modify tensorboard callback
